+++ pkg/ccl/changefeedccl/BUILD.bazel
@@ -25,6 +25,7 @@ go_library(
         "//pkg/ccl/backupccl/backupresolver",
         "//pkg/ccl/changefeedccl/changefeedbase",
         "//pkg/ccl/changefeedccl/changefeeddist",
+        "//pkg/ccl/changefeedccl/kvevent",
         "//pkg/ccl/changefeedccl/kvfeed",
         "//pkg/ccl/utilccl",
         "//pkg/clusterversion",
@@ -121,6 +122,7 @@ go_test(
         "//pkg/blobs",
         "//pkg/ccl/changefeedccl/cdctest",
         "//pkg/ccl/changefeedccl/changefeedbase",
+        "//pkg/ccl/changefeedccl/kvevent",
         "//pkg/ccl/changefeedccl/kvfeed",
         "//pkg/ccl/importccl",
         "//pkg/ccl/storageccl",
