+++ pkg/ccl/changefeedccl/kvfeed/BUILD.bazel
@@ -3,9 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "kvfeed",
     srcs = [
-        "buffer.go",
         "kv_feed.go",
-        "metrics.go",
         "physical_kv_feed.go",
         "scanner.go",
     ],
@@ -13,6 +11,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/ccl/changefeedccl/changefeedbase",
+        "//pkg/ccl/changefeedccl/kvevent",
         "//pkg/ccl/changefeedccl/schemafeed",
         "//pkg/gossip",
         "//pkg/jobs/jobspb",
@@ -22,22 +21,14 @@ go_library(
         "//pkg/kv/kvserver",
         "//pkg/roachpb",
         "//pkg/settings/cluster",
-        "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/lease",
         "//pkg/sql/covering",
-        "//pkg/sql/rowcontainer",
-        "//pkg/sql/rowenc",
-        "//pkg/sql/sem/tree",
-        "//pkg/sql/types",
         "//pkg/storage/enginepb",
         "//pkg/util/ctxgroup",
-        "//pkg/util/envutil",
         "//pkg/util/hlc",
         "//pkg/util/log",
-        "//pkg/util/metric",
         "//pkg/util/mon",
         "//pkg/util/span",
-        "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
     ],
@@ -50,6 +41,7 @@ go_test(
     embed = [":kvfeed"],
     deps = [
         "//pkg/ccl/changefeedccl/changefeedbase",
+        "//pkg/ccl/changefeedccl/kvevent",
         "//pkg/ccl/changefeedccl/schemafeed",
         "//pkg/ccl/changefeedccl/schemafeed/schematestutils",
         "//pkg/keys",
